+++ Squickly_Monsters/homepage/interactions.lua
@@ -249,6 +249,7 @@ end
 function mostRecentPlay2Clicked(event)
     if isTouchAble then
         if event.phase == "ended" then
+            print("mostRecentPlay2Clicked")
             hideShowAllIcons(playIconsList)
             changeToWakeupState()
             playAnimation()
