+++ Squickly_Monsters/inventory/interactions.lua
@@ -7,12 +7,14 @@ local maxSize = 9;
 
 -- -------------------------------------------------------------------------------
 --functions for currency
-local gold --default goldMoney
-local platinum--default platinumMoney
+local gold=0 --default goldMoney
+local platinum=0--default platinumMoney
 
+-- Used for changing the amount of Gold and Platinum user has. 
 function updateCurrency(goldCost, platinumCost)
 	gold = gold + goldCost
 	platinum = platinum + platinumCost
+	print("GOld" .. gold)
 end
 
 function sufficientGold(goldCost)
