+++ Squickly_Monsters/shop/background.lua
@@ -1,4 +1,4 @@
--------------------------------------------------------------------------------
+-- -----------------------------------------------------------------------------
 -- Local variables go HERE
 
 local backgroundShop;
